:recycle: 마이페이지 SSR 기준으로 코드 수정
