+++ package.json
@@ -21,6 +21,7 @@
     "axios": "^1.6.8",
     "moment": "^2.30.1",
     "next": "14.2.2",
+    "next-auth": "^4.24.7",
     "react": "^18",
     "react-calendar": "^5.0.0",
     "react-dom": "^18",
