+++ src/app/(auth)/mypage/page.tsx
@@ -1,73 +1,19 @@
-"use client"
-import * as S from "@/app/(auth)/mypage/mypage.style";
-import Nickname from "@/app/components/mypage/Form/NickName/Nickname";
-import Password from "@/app/components/mypage/Form/Password/Password";
-import UserInfo from "@/app/components/mypage/UserInfo/UserInfo";
-import { MypageType } from "@/app/types/mypage";
-import { useAppSelector } from "@/store/hook";
-import axios from "axios";
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import { MoonLoader } from "react-spinners";
 
-export default function Mypage() {
+import MypageBase from "@/app/components/mypage/MypageBase/MypageBase";
+import { cookies } from "next/headers";
 
-  const router = useRouter();
-  const user = useAppSelector(state=>state.user);
-  const [mypage,setMypage] = useState<MypageType | null>(null);
+export default async function Mypage() {
 
-  const userDeleteHanlder = ()=>{
-    alert('회원탈퇴 기능은 준비중입니다.');
-  }
-
-  // 회원가져오기
-  useEffect(()=>{
-
-    axios.get('/api/member/mypage',{
-      headers : {
-        Authorization : user
-      }
-    })
-    .then((res)=>{
-      const {status,message,data} = res.data;
-
-      if(status === "success"){
-        setMypage(data);
-      }else{
-        alert(message);
-      }
-
-    })
-    .catch(()=>{
-      console.log('서버 에러');
-    })
-
-  },[user]);
+  const token = cookies().get('user')?.value;
+  if(!token) return;
+  const res = await fetch(`${process.env.API_URL}/api/v1/member`,{
+    headers : {
+      Authorization : token
+    }
+  });
+  const {status,message,data} = await res.json();
 
   return (
-    <>
-      {
-        !mypage 
-        ? 
-          <div style={{top : 0, left : 0, width : "100%", height : "100%", display : "flex", alignItems : "center", justifyContent : "center", background : "rgba(255,255,255,0.2)", backdropFilter : "blur(10px)"}}>
-            <MoonLoader color="#574fff"></MoonLoader>
-          </div>
-        :
-        <S.MypageStyle>
-          <S.Wrap>
-            <S.Prev onClick={()=>router.back()}><img src="/asset/icon/prev_arrow.svg" alt="뒤로가기" /></S.Prev>
-            <S.Title>마이페이지</S.Title>
-            <UserInfo mypage={mypage}/>
-          </S.Wrap>
-          <S.Wrap>
-            <Nickname setMypage={setMypage}/>
-            <Password/>
-            <S.UserDelete>
-                <button onClick={userDeleteHanlder}>회원탈퇴하기</button>
-            </S.UserDelete>
-          </S.Wrap>
-        </S.MypageStyle>
-      }
-    </>
+    <MypageBase data={data}/>
   )
 }
