+++ src/app/(route)/del/[postid]/detail.style.ts
