+++ src/app/(route)/del/[postid]/loading.tsx
