+++ src/app/api/post/get/route.ts
@@ -1,11 +1,12 @@
 import axios from "axios";
-import { headers } from "next/headers";
+import { cookies } from "next/headers";
 import { NextRequest, NextResponse } from "next/server"
 
 export const GET = async (request : NextRequest)=>{
 
-    const headerList = headers();
-    const token = headerList.get("authorization");
+    if(!process.env.API_URL) throw new Error('env 에러가 발생했습니다.');
+
+    const token = cookies().get('user')?.value;
 
     const searchParams = request.nextUrl.searchParams;
     const page = searchParams.get('page');
@@ -17,25 +18,19 @@ export const GET = async (request : NextRequest)=>{
 
     if(!process.env.API_URL) throw new Error('env 에러가 발생했습니다.');
 
-    try {
-        const response = await axios.get(`${process.env.API_URL}/api/v1/meeting`,{
-            params :{
-                page,
-                meetingstatus,
-                q,
-                location,
-                gametype,
-                sort
-            },
-            headers : {
-                Authorization : token
-            }
-        });
-        return new NextResponse(JSON.stringify(response.data));
-    }
-    catch(err){
-        console.log(err);
-        return new NextResponse('서버 에러');
-    }
+    const response = await axios.get(`${process.env.API_URL}/api/v1/meeting`,{
+        params :{
+            page,
+            meetingstatus,
+            q,
+            location,
+            gametype,
+            sort
+        },
+        headers : {
+            Authorization : token
+        }
+    });
+    return new NextResponse(JSON.stringify(response.data));
 
 }
