+++ src/app/api/post/my/route.ts
@@ -1,17 +1,16 @@
 import axios from "axios";
-import { headers } from "next/headers"
+import { cookies } from "next/headers"
 import { NextRequest, NextResponse } from "next/server";
 
 export const GET = async (req : NextRequest)=>{
 
+    if(!process.env.API_URL) throw new Error('env 에러가 발생했습니다.');
+
+    const token = cookies().get('user')?.value;
+
     const searchParams = req.nextUrl.searchParams;
     const page = searchParams.get('page');
     const q = searchParams.get('q');
-
-    const headerList = headers();
-    const token = headerList.get("authorization");
-
-    if(!process.env.API_URL) throw new Error('env 에러가 발생했습니다.');
 
     try {
         const response = await axios.get(`${process.env.API_URL}/api/v1/meeting/my`,{
