+++ src/app/components/common/Mypage/Change/Change.tsx
@@ -3,6 +3,6 @@ import {ChangeType} from "./Change.stories";
 
 export default function Change({onClick,children} : ChangeType) {
   return (
-    <S.Button onClick={onClick}>{children}</S.Button>
+    <S.Button type="button" onClick={onClick}>{children}</S.Button>
   )
 }
